feat(gateway): move builder inside gateway folder
