+++ lib/gateway/build-gateway.js
@@ -7,18 +7,18 @@ const {
   Kind
 } = require('graphql')
 const { Factory } = require('single-user-cache')
-const buildFederatedSchema = require('./federation')
-const buildServiceMap = require('./gateway/service-map')
+const buildFederatedSchema = require('../federation')
+const buildServiceMap = require('./service-map')
 const {
   makeResolver,
   createQueryOperation,
   createFieldResolverOperation,
   createEntityReferenceResolverOperation,
   kEntityResolvers
-} = require('./gateway/make-resolver')
-const { MER_ERR_GQL_GATEWAY_REFRESH, MER_ERR_GQL_GATEWAY_INIT, MER_ERR_SERVICE_RETRY_FAILED } = require('./errors')
-const findValueTypes = require('./gateway/find-value-types')
-const getQueryResult = require('./gateway/get-query-result')
+} = require('./make-resolver')
+const { MER_ERR_GQL_GATEWAY_REFRESH, MER_ERR_GQL_GATEWAY_INIT, MER_ERR_SERVICE_RETRY_FAILED } = require('../errors')
+const findValueTypes = require('./find-value-types')
+const getQueryResult = require('./get-query-result')
 
 function isDefaultType (type) {
   return [
